requirements.txt
pytesseract
rapidocr-onnxruntime
opencv-python-headless
fastapi
uvicorn
python-multipart
numpy
requests
pillow
httpx